only depend on hibernate-validator in test scope
typically, a validator is already provided by an app
server. It should be possible to use another validator
than the reference implementation (depend on APIs,
not on implementations).

Upgrade hibernate-validator to 5.4.3 due to vulnerabilities.
Unfortunately, upgrading to later version is not possible,
because that would require BeanValidation 2.0 API, which
are not yet supported in older projects / app servers.